+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.1 (2021-07-07)
+
++ Added `BaseSerializer` _(DRF)_ and `BaseModel` _(pydantic)_ support for `JsonCodec.encode`
++ Added `djantic` _(Django models for `pydatic`)_ models support
+
 ## 0.3.0 (2021-07-07)
 
 + Added `pydantic` models support